refactor(account): split privacy fax dialog
